+++ Binaries/dcmtk-source/dcmdata/dcvrus.cc
@@ -116,7 +116,7 @@ void DcmUnsignedShort::print(ostream &out,
                 else
                     sprintf(buffer, "\\%hu", *uintVals);
                 /* check whether current value sticks to the length limit */
-                newLength = printedLength + strlen(buffer);
+                newLength = printedLength + (unsigned int)strlen(buffer);
                 if ((newLength <= maxLength) && ((i + 1 == count) || (newLength + 3 <= maxLength)))
                 {
                     out << buffer;
